add support for Anker Solix C300X portable power station
